Create authorization system for update / delete

+++ app/Http/Controllers/Pesantren/DeleteController.php
@@ -8,27 +8,47 @@ use App\Http\Controllers\Controller;
 use App\Pesantren;
 use App\Potential;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class DeleteController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function hapusPesantren($id){
-        $fotopesantren = FotoPesantren::where('pesantrenid', $id)->get();
-        foreach ($fotopesantren as $foto) {
+        $pesantren = Pesantren::where('id', $id);
+        if (Auth::id() != $pesantren->first()->pembuatid) {
+            return redirect()->back();
+        }
+
+        $fotopesantren = FotoPesantren::where('pesantrenid', $id);
+        foreach ($fotopesantren->get() as $foto) {
             Storage::delete($foto->img);
         }
 
         DB::table('potensi_pesantren')->where('pesantrenid', $id)->delete();
-        FotoPesantren::where('pesantrenid', $id)->delete();
-        Pesantren::where('id', $id)->delete();
+        $fotopesantren->delete();
+        $pesantren->delete();
 
         return redirect()->back();
     }
 
     public function hapusFoto($id){
-        $foto = FotoPesantren::where('id', $id)->first();
-        Storage::delete($foto->img);
-        FotoPesantren::where('id', $id)->delete();
+        $foto = FotoPesantren::where('id', $id);
+        if (Auth::id() != $foto->first()->pembuatid) {
+            return redirect()->back();
+        }
+
+        Storage::delete($foto->first()->img);
+        $foto->delete();
 
         return redirect()->back();
     }
